Minor review changes
